Move juttle-moment.js to lib/runtime/types
