+++ lib/moment/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var JuttleMoment = require('./juttle-moment');
+var JuttleMoment = require('../runtime/types/juttle-moment');
 
 module.exports = {
     JuttleMoment: JuttleMoment
