+++ test/runtime/types/humanized-moments.spec.js
@@ -1,10 +1,10 @@
 'use strict';
 
 var _ = require('underscore');
-var juttle_test_utils = require('./specs/juttle-test-utils');
+var juttle_test_utils = require('../specs/juttle-test-utils');
 var check_juttle = juttle_test_utils.check_juttle;
 var expect = require('chai').expect;
-var JuttleMoment = require('../../lib/moment').JuttleMoment;
+var JuttleMoment = require('../../../lib/moment').JuttleMoment;
 
 var validDurationUnits = [
     'week',
